requirements.txt
PyMuPDF~=1.24.10
elasticsearch~=8.15.1
python-dotenv~=1.0.1